Change jtag command for reading flash
Allows tockloader to read attributes from a generic device in order to
determine what the actual device is. Previously, the reset command in
JTAG would sometimes fail when run as a generic `cortex-m0` device
(rather than a specific `ATSAM4LC8C` for instance). However, reading
flash still works, even if reset fails, so we now just do that if
running on a generic device.