Fix a bug in UTF decoding if no bootloader present
Checking for a bootloader tries to decode the UTF-8 string
"TOCKBOOTLOADER" at a particular address. However, if this address
doesn't contain a bootloader, it _may_ contain a non-UTF8 decodable
byte.

We've been using implicit 'strict' error handling for the UTF package,
which throws an exception.  This change simply changes that to 'ignore'.