Fixed input message
There was a problem with the message, if a user
entered a wrong number. The message (with default
number) keeps getting bigger.

Also there was a problem if the user selected the
"last" element. If there were 4 possible ports like
[0] first one
[1] second one
[2] third one
[3] fourth one
and the user entered "4" the program crashed because
the program accepted 4 (the length of the array)
as valid input.

Both errors are removed!