I'm able to install two apps now
The erase command right after write_image actually erases the entire
SECTOR. And I could not find a way to make it write correctly without
it.

On the other hand, all apps are installed now, but only the first one
gets executed. So that's probably a bug in another part of the code.