remove _reflash_app_headers
This was an optimization because if a user is only setting a flag in an
app's header, then we don't need to re-flash the entire app binary.
However, as tockloader becomes more complex, having two app updating
functions is difficult to maintain. While this will be a _bit_ slower,
setting flags is relatively rare so in practice this won't matter much.

This consolidates everything to reshuffle apps. Also, it makes the
notion of a modified app more explicit, since a modified app must be
re-written to the board.