Add a flash_file board interface to manipulate board flash dumps
Adds a new board interface used to manipulate board flash dumps. This
can be useful for building an image containing a kernel and apps for
flashing with other tools, for instance with proprietary toolchains or
booting an image via TFTP. Furthermore, this can be useful for
building images which are embedded into FPGA bitstreams.

Signed-off-by: Leon Schuermann <leon@is.currently.online>