openocd: Fix cleanup after exceptions
This fixes unwinding resources after an exception. The old version would exit if the creation of the socket failed, raising an exception when trying to close the socket.

The new version makes sure that only objects successfully created are cleaned up.

Instead of nesting try..finally multiple times, this appends handlers onto an explicit list.