TabApp.app_has_binary: reflect empty TBF list, check during install
In case of a missing binary compiled for a certain architecture,
missspelled architecture name, or an otherwise invalid TBF, it is
possible for an AppTbf to be constructed over an empty TBF list. In
this case, this should be reflected in the result of the
`app_has_binary` method. Furthermore, Tockloader should verify that
`extract_app` yields a `TabApp` for which `app_has_binary` is true.

Without this change, when trying to install an app for an architecture
where no compiled binary exists, an error such as the following is
encountered:

    [STATUS ] Installing app on the board...
    [ERROR  ] Size only valid with one TBF

This error message is misleading, as it occurs only when shuffling
apps in the PIC code path, as no app requring fixed address placement
is found, and also has no apparent relation to an architecture's app
binaries not being found. Now, a more helpful message is shown:

    [STATUS ] Installing app on the board...
    [ERROR  ] Unable to locate a valid application binary matching the target architecture (rv32i)

Signed-off-by: Leon Schuermann <leon@is.currently.online>