Autodetect the stm32f3x and stm32f4x from the output of lsusb
With this change, it becomes possible to run the "getting started"
guide on these boards by just adding --openocd flag as advised in the guide.

It was successfully tested with a STM32 F44RE nucleo board. I'm still
using the stm32f4discovery board name from the config because it's the
one working for me. I failed to use the 'nucleof4' board from that set.

This commit also makes an error message more explicit.