openocd: finer-grained error msg if board unknown
When trying to set up tockloader to use openocd, I kept getting
the generic message saying "Could not determine the current board
or arch or openocd board name". Not only did it not mention missing
page_size parameter, but also was too coarse-grained. With splitting
the error message, I managed to debug this, so I believe this could be
beneficial to anyone.